+++ CHANGES.md
@@ -17,11 +17,14 @@ Release Notes.
 * Add Seata in the component definition. Seata plugin hosts on Seata project.
 * Extended Kafka plugin to properly trace consumers that have topic partitions directly assigned.
 * Support print SkyWalking context to logs.
+* Add `MessageListener` enhancement in pulsar plugin
 
 #### OAP-Backend
 * BugFix: filter invalid Envoy access logs whose socket address is empty.
 * Fix K8s monitoring the incorrect metrics calculate. 
 * Loop alarm into event system.
+* Support alarm tags.
+* Support WeLink as a channel of alarm notification.
 
 #### UI
 * Add logo for kong plugin.